#4924 Gestion des Projets: gestion de projet - suivi de commande
- https://suivi.espai-web.com/tache/4924
- fix event position

+++ assets/js/angular/scheduler/px.service.js
@@ -1,9 +1,9 @@
 angular.module('schedulerModule').factory('px', [function() {
     return (value) => {
-        if (isNaN(parseInt(value))) {
+        if (isNaN(parseFloat(value))) {
             return value;
         }
 
-        return parseInt(value) + 'px';
+        return parseFloat(value) + 'px';
     }
 }]);
